Extract Crawl to a library
Also adds a test and a benchmark
